CLients/GUI/requirements.txt
attrs==22.2.0
Babel==2.10.3
bcrypt==3.2.2
blinker==1.5
# Brlapi==0.8.4
certifi==2022.9.24
chardet==5.1.0
click==8.1.3
# cloud-init==23.3.3
# colorama==0.4.6
# command-not-found==0.3
configobj==5.0.8
cryptography==38.0.4
# cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
distro==1.8.0
distro-info==1.5+ubuntu0.23.4.1
duplicity==0.8.22
fasteners==0.17.3
future==0.18.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.6.0
keyring==23.9.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.24.0
Mako==1.2.4.dev0
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mysql-connector-python==8.0.15
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
paramiko==2.12.0
pexpect==4.8.0
Pillow==9.4.0
protobuf==4.21.12
ptyprocess==0.7.0
PyAudio==0.2.13
pycairo==1.20.1
pycups==2.0.1
Pygments==2.14.0
PyGObject==3.44.1
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.5.3+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
pytz==2022.7.1
pyxdg==0.28
PyYAML==6.0
reportlab==3.6.12
requests==2.28.1
rich==13.3.1
SecretStorage==3.3.3
six==1.16.0
systemd-python==235
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.12
usb-creator==0.3.16
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
